+++ github/packages.py
@@ -3,6 +3,7 @@ import logging
 import re
 import urllib.parse
 from http import HTTPStatus
+from typing import TypeVar
 
 import github_action_utils as gha_utils
 
@@ -13,6 +14,7 @@ from github.models.packages.user import Package as UserPackage
 from utils.errors import RateLimitError
 
 logger = logging.getLogger(__name__)
+T = TypeVar("T")
 
 
 class ContainerPackage(GithubEndpointResponse):
@@ -64,7 +66,7 @@ class ContainerPackage(GithubEndpointResponse):
         return f"Package {self.name}"
 
 
-class _GithubContainerRegistryApiBase(GithubApiBase):
+class _GithubContainerRegistryApiBase[T](GithubApiBase[T]):
     PACKAGE_VERSIONS_ENDPOINT: str = ""
     PACKAGE_VERSION_DELETE_ENDPOINT: str = ""
     PACKAGE_VERSION_RESTORE_ENDPOINT: str = ""
@@ -106,8 +108,8 @@ class _GithubContainerRegistryApiBase(GithubApiBase):
 
         pkgs = []
 
-        for data in await self._read_all_pages(endpoint, query_params=query_params):
-            pkgs.append(ContainerPackage(data))
+        for data in await self.list(endpoint, query_params=query_params):
+            pkgs.append(ContainerPackage(data))  # type: ignore[arg-type]
 
         return pkgs
 
@@ -123,24 +125,11 @@ class _GithubContainerRegistryApiBase(GithubApiBase):
     ) -> list[ContainerPackage]:
         return await self.versions(package_name, active=False)
 
-    async def delete(self, package_data: ContainerPackage):
+    async def delete_package(self, package_data: ContainerPackage):
         """
         Deletes the given package version from the GHCR
         """
-        resp = await self._client.delete(package_data.url)
-        if resp.status_code != HTTPStatus.NO_CONTENT:
-            # If forbidden, check if it is rate limiting
-            if resp.status_code == HTTPStatus.FORBIDDEN and "X-RateLimit-Remaining" in resp.headers:
-                remaining = int(resp.headers["X-RateLimit-Remaining"])
-                if remaining <= 0:
-                    raise RateLimitError
-            else:
-                msg = f"Request to delete {package_data.url} returned HTTP {resp.status_code}"
-                gha_utils.warning(
-                    message=msg,
-                    title=f"Unexpected delete status: {resp.status_code}",
-                )
-                logger.warning(msg)
+        await self.delete(package_data.url)
 
     async def restore(
         self,
@@ -171,7 +160,7 @@ class _GithubContainerRegistryApiBase(GithubApiBase):
                 logger.warning(msg)
 
 
-class GithubContainerRegistryOrgApi(_GithubContainerRegistryApiBase):
+class GithubContainerRegistryOrgApi(_GithubContainerRegistryApiBase[OrgPackage]):
     """
     Class wrapper to deal with the GitHub packages API.  This class only deals with
     container type packages.
@@ -187,7 +176,7 @@ class GithubContainerRegistryOrgApi(_GithubContainerRegistryApiBase):
     )
 
 
-class GithubContainerRegistryUserApi(_GithubContainerRegistryApiBase):
+class GithubContainerRegistryUserApi(_GithubContainerRegistryApiBase[UserPackage]):
     """
     Class wrapper to deal with the GitHub packages API.  This class only deals with
     container type packages.